Create printer-prusa-mk3s-2021.cfg
Example for Prusa MK3S and MK3. Based largely on my MK3S Bear config. 
It includes macros (I noticed most other examples don't) but some of these are needed for the filament sensor as well as to restore the Tram Z/ Calibrate Z functionality. I also really wanted something that would work fully out-of-the-box even with a UI like Mainsail.   
Explicitly put the Prusa USB-to-serial issue note in there as well as workarounds based on feedback by Arksine.